refactoring packages of cards, modified navigation to not show splash page
