+++ app/src/main/java/com/example/sampleappbyme/main/ViewModelFactory.kt
@@ -4,7 +4,6 @@ import android.annotation.SuppressLint
 import android.app.Application
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.ViewModelProviders
 import com.example.sampleappbyme.main.data.SampleRepository
 import com.example.sampleappbyme.main.viewmodel.MainViewModel
