+++ app/src/main/java/com/example/sampleappbyme/main/ui/card/CardFragmentPagerAdapter.kt
@@ -1,9 +1,11 @@
-package com.example.sampleappbyme.main.ui
+package com.example.sampleappbyme.main.ui.card
 
 import androidx.cardview.widget.CardView
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentStatePagerAdapter
+import com.example.sampleappbyme.main.ui.card.CardAdapter
+import com.example.sampleappbyme.main.ui.card.CardFragment
 
 class CardFragmentPagerAdapter : CardAdapter, FragmentStatePagerAdapter {
