+++ app/src/main/res/navigation/nav_graph.xml
@@ -15,7 +15,6 @@
             android:id="@+id/action_go_to_recycler_view_fragment"
             app:destination="@id/recyclerViewFragment"
             app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@anim/slide_out_right"
             app:popExitAnim="@anim/slide_in_left"
             app:popEnterAnim="@anim/slide_in_right"/>
 
@@ -23,9 +22,7 @@
             android:id="@+id/action_go_to_card_view_fragment"
             app:destination="@id/cardViewFragment"
             app:enterAnim="@anim/slide_in_right"
-            app:exitAnim="@anim/slide_out_right"
-            app:popExitAnim="@anim/slide_in_left"
-            app:popEnterAnim="@anim/slide_in_right"/>
+            app:popExitAnim="@anim/nav_default_exit_anim"/>
     </fragment>
 
     <fragment
@@ -39,6 +36,8 @@
             app:destination="@id/mainMenuFragment"
             app:enterAnim="@anim/slide_in_right"
             app:exitAnim="@anim/slide_out_right"
+            app:popUpTo="@id/nav_graph"
+            app:popUpToInclusive="true"
             app:popExitAnim="@anim/slide_in_left"
             app:popEnterAnim="@anim/slide_in_right"/>
     </fragment>
@@ -47,13 +46,13 @@
         android:id="@+id/recyclerViewFragment"
         android:name="com.example.sampleappbyme.main.view.fragment.RecyclerViewFragment"
         android:label="activity_main"
-        tools:layout="@layout/main_fragment">
+        tools:layout="@layout/recyclerview_fragment">
     </fragment>
 
     <fragment
         android:id="@+id/cardViewFragment"
         android:name="com.example.sampleappbyme.main.view.fragment.CardViewFragment"
         android:label="activity_main"
-        tools:layout="@layout/main_fragment">
+        tools:layout="@layout/card_view_fragment">
     </fragment>
 </navigation>
